Rename uiParams to parametersDefinition, Move schema based methods to UniversalSchemaSupportDispatcher
