+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/source/UniversalKafkaSourceFactory.scala
@@ -1,15 +1,13 @@
 package pl.touk.nussknacker.engine.schemedkafka.source
 
-import cats.data.{NonEmptyList, Validated, ValidatedNel}
+import cats.data.{NonEmptyList, Validated}
 import cats.data.Validated.Valid
 import io.circe.Json
 import io.circe.syntax._
 import io.confluent.kafka.schemaregistry.ParsedSchema
-import io.confluent.kafka.schemaregistry.avro.AvroSchema
-import io.confluent.kafka.schemaregistry.json.JsonSchema
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.apache.kafka.common.record.TimestampType
-import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.{CustomNodeError, FatalUnknownError}
+import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.CustomNodeError
 import pl.touk.nussknacker.engine.api.context.transformation.{DefinedEagerParameter, NodeDependencyValue}
 import pl.touk.nussknacker.engine.api.context.{ProcessCompilationError, ValidationContext}
 import pl.touk.nussknacker.engine.api.definition._
@@ -17,20 +15,14 @@ import pl.touk.nussknacker.engine.api.process.{ContextInitializer, ProcessObject
 import pl.touk.nussknacker.engine.api.test.TestRecord
 import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypedClass, TypingResult, Unknown}
 import pl.touk.nussknacker.engine.api.util.NotNothing
-import pl.touk.nussknacker.engine.api.validation.ValidationMode
 import pl.touk.nussknacker.engine.api.{MetaData, NodeId}
-import pl.touk.nussknacker.engine.json.JsonSinkValueParameter
-import pl.touk.nussknacker.engine.kafka.KafkaFactory.SinkValueParamName
 import pl.touk.nussknacker.engine.kafka.PreparedKafkaTopic
 import pl.touk.nussknacker.engine.kafka.consumerrecord.SerializableConsumerRecord
 import pl.touk.nussknacker.engine.kafka.source._
 import pl.touk.nussknacker.engine.kafka.source.KafkaSourceFactory.{KafkaSourceImplFactory, KafkaTestParametersInfo}
 import pl.touk.nussknacker.engine.schemedkafka.KafkaUniversalComponentTransformer.SchemaVersionParamName
-import pl.touk.nussknacker.engine.schemedkafka.encode.BestEffortAvroEncoder
 import pl.touk.nussknacker.engine.schemedkafka.schemaregistry._
 import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.formatter.SchemaBasedSerializableConsumerRecord
-import pl.touk.nussknacker.engine.schemedkafka.schemaregistry.universal.{AvroSchemaSupport, JsonPayloadRecordFormatterSupport, UnsupportedSchemaType}
-import pl.touk.nussknacker.engine.schemedkafka.sink.AvroSinkValueParameter
 import pl.touk.nussknacker.engine.schemedkafka.source.UniversalKafkaSourceFactory.UniversalKafkaSourceFactoryState
 import pl.touk.nussknacker.engine.schemedkafka.{KafkaUniversalComponentTransformer, RuntimeSchemaData}
 
@@ -127,7 +119,6 @@ class UniversalKafkaSourceFactory[K: ClassTag: NotNothing, V: ClassTag: NotNothi
     implicit val nodeId: NodeId = TypedNodeDependency[NodeId].extract(dependencies)
 
     val preparedTopic = extractPreparedTopic(params)
-    val versionOption = parseVersionOption(params)
     val UniversalKafkaSourceFactoryState(keySchemaDataUsedInRuntime, valueSchemaUsedInRuntime, kafkaContextInitializer) = finalState.get
 
     // prepare KafkaDeserializationSchema based on given key and value schema (with schema evolution)
@@ -138,41 +129,24 @@ class UniversalKafkaSourceFactory[K: ClassTag: NotNothing, V: ClassTag: NotNothi
     val formatterSchema = schemaBasedMessagesSerdeProvider.deserializationSchemaFactory.create[K, V](kafkaConfig, None, None)
     val recordFormatter = schemaBasedMessagesSerdeProvider.recordFormatterFactory.create[K, V](kafkaConfig, formatterSchema)
     implProvider.createSource(params, dependencies, finalState.get, List(preparedTopic), kafkaConfig, deserializationSchema,
-      recordFormatter, kafkaContextInitializer, getParameters(valueSchemaUsedInRuntime, preparedTopic.original))
+      recordFormatter, kafkaContextInitializer, prepareKafkaTestParametersInfo(valueSchemaUsedInRuntime, preparedTopic.original))
   }
 
-  private def getParameters(runtimeSchema: Option[RuntimeSchemaData[ParsedSchema]], topic: String)
+  private def prepareKafkaTestParametersInfo(runtimeSchema: Option[RuntimeSchemaData[ParsedSchema]], topic: String)
                            (implicit nodeId: NodeId): KafkaTestParametersInfo = {
     Validated.fromOption(runtimeSchema, NonEmptyList.one(CustomNodeError(nodeId.id, "Cannot generate test parameters: no runtime schema found", None)))
-      .andThen(handleSchemaType(_, topic))
-      .valueOr(e => throw new RuntimeException(e.toList.mkString("")))
-  }
-
-  private def handleSchemaType(determinedSchema: RuntimeSchemaData[ParsedSchema], topic: String)(implicit nodeId: NodeId): ValidatedNel[ProcessCompilationError, KafkaTestParametersInfo] = {
-    (determinedSchema.schema match {
-      case s: AvroSchema => AvroSinkValueParameter(s.rawSchema()).map(_.toParameters)
-      case s: JsonSchema => JsonSinkValueParameter(s.rawSchema(), SinkValueParamName, ValidationMode.lax)(nodeId).map(_.toParameters)
-      case s => Validated.invalidNel(FatalUnknownError(s"Avro or Json schema is required, but got ${s.schemaType()}"))
-    }).map { params => KafkaTestParametersInfo(params, prepareTestRecord(determinedSchema, topic)) }
+      .andThen { schema => schemaSupportDispatcher
+        .parametersFromRuntimeParsedSchema(schema)
+        .map { params => KafkaTestParametersInfo(params, prepareTestRecord(schema, topic)) }
+      }.valueOr(e => throw new RuntimeException(e.toList.mkString("")))
   }
 
   private def prepareTestRecord(schema: RuntimeSchemaData[ParsedSchema], topic: String): Any => TestRecord = any => {
-    val json = prepareMessageFormatter(schema.schema)(any)
+    val json = schemaSupportDispatcher.messageFormatterFromParsedSchema(schema.schema, schemaRegistryClient)(any)
     val serializedConsumerRecord = SerializableConsumerRecord[Json, Json](None, json, Some(topic), None, None, None, None, None, None)
     TestRecord(
       SchemaBasedSerializableConsumerRecord[Json, Json](None, schema.schemaIdOpt, serializedConsumerRecord).asJson
     )
-  }
-
-  private def prepareMessageFormatter(schema: ParsedSchema): Any => Json = {
-    schema match {
-      case s: AvroSchema =>
-        val encoder = BestEffortAvroEncoder(ValidationMode.lax)
-        val recordFormatterSupport = new AvroSchemaSupport(kafkaConfig).recordFormatterSupport(schemaRegistryClient)
-        (data: Any) => recordFormatterSupport.formatMessage(encoder.encodeOrError(data, s.rawSchema()))
-      case _: JsonSchema => JsonPayloadRecordFormatterSupport.formatMessage
-      case _ => throw new UnsupportedSchemaType(schema.schemaType())
-    }
   }
 
   override def nodeDependencies: List[NodeDependency] = List(TypedNodeDependency[MetaData],
